requirements.txt
tensorflow>=2.13
matplotlib>=3.7
numpy>=1.25
Pillow>=10.0
pandas>=2.1
